backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
python-multipart==0.0.6
pydantic==2.5.3
langchain==0.1.0
langchain-openai==0.0.2
langchain-community==0.0.10
openai==1.10.0
pypdf==3.17.4
faiss-cpu==1.7.4
numpy==1.24.3
tiktoken==0.5.2
aiofiles==23.2.1
pandas==2.1.4
